rename `sleep` module to `power`
