+++ src/sysc/mod.rs
@@ -7,7 +7,7 @@ pub mod ledctl;
 pub mod logging;
 pub mod net;
 pub mod ota;
-pub mod sleep;
+pub mod power;
 pub mod usbctl;
 
 pub use error::{OsError, ReportableError};
